requirements.txt
assemblyai
google-generativeai
python-dotenv